Test: Fix Issue After Mergo of #5820
Ok, I introduced an issue, when I merged
https://github.com/ILIAS-eLearning/ILIAS/pull/5820 as the
first pass is pass 0. The function _getBestPass() didn't
act as it should and returned 0 if no pass existed as well
as when the first pass, was the one we wanted. Fixed with
making the pass check stricter and by making_getBestPass()
return null, when there is no best pass for whatever reason.