UI: change UploadHandler to fix #37161
@klees: Excerpt from discussion on GitHub:

The analysis in the ticket lists the steps that lead to the problem:

* file-input receives existing file-id (by ::withValue())
* user removes the existing file-id to upload a new one
* file-input deletes the old file before final submission
* file-input uploads the new file before final submission
* user submits
* $_POST receives new (existing) file-id)
* server builds the same form again, incl. ::withValue() call which is now a
  file-id that does not exist anymore.
* file-input receives invalid file-id, due to missing validation
  (UploadHandler::getInfoResult()) the id will be added as dynamic input
* ::withInput() is called because this is a submission
* ::withValue() will be called again with the new POST data, which also contains
  a file-id, the existing one now.

The problem is this: We already delete the file asynchronously even before the user
even confirmed that she really wants to have this file instead of that file by
submitting the form. This also is problematic in other cases: When user uses [Cancel]
or browsers [Back]. There might be more...

This is not a problem when we are adding files. Because then we only have an unused
file lying around if the user decides to [Cancel]. But deleted files are just gone.
Which is exactly what this PR would be hiding. This would not hide the [Cancel] or
[Back]-problem, though.

So the root cause is that we even want to getFileRemovalURL from our users. We cannot
really use that URL in a sensible way IMO, because we will only know that a user
does not want a file anymore once the POST hits the server. Also note, that even then
we, as a UI framework, cannot know that we really can delete the file. Might be that
we need to still keep it, because it is already used somewhere else. Hence: IRSS.

So, the fix from the perspective of the UI framework is: remove getFileRemovalURL
from the interface and do not delete files asynchronously.

@tbongers-cat:
Removed all implementations of getFileRemoval() and getRemoveResult() as well as the usage in Renderer/JS
Marked getRemoveResult as final and adapted the README