TA: patch for #37559
This does not fix the root cause, though. The id should never be a string in the
first place, hence we should tackle the form that causes this at some point.
Obvious problem would be, that a missing value in the field would do something
with something with id=0, because (int)null === 0, sadly.