[Mantis #37524] Ignore JavaScript links in URLs
Before the patch, the test player would submit JavaScript
URLs (javascript:) to the server as a redirect target.
The browser would follow the redirect and get a 404 from
the web server.

e.g. User clicks on volume button linking to
javascript:void(0), the test player would add a form field
containing javascript:void(0), the server would send a
redirect to /javascript:void(0); the browser requests that
page, the server sends a 404 Not Found error.

https://mantis.ilias.de/view.php?id=37524