Test: Fix Some PHP8.1 Errors + Clean-Up
This adds some inline JS in ilTestNavigationToolbarGUI. I decided to do it
this way, as otherwise I would have needed to create a new JS file and then
to add it to the page. With the current JS structure this seemed confusing
and hard to read to me. One more sign that we need to restructure JS.