restructure directories
This implements https://github.com/ILIAS-eLearning/ILIAS/blob/trunk/docs/development/components-and-directories.md#filesystem-and-namespaces.
---
Hello everyone,

the trunk is open again, after we have moved a lot of files according to [1][2]. Thanks Luka Stocker!

Please make sure to understand that we have moved all resources and endpoints to the `/public`-folder, so you need to point your webserver or browser there (instead of to the root of our directory tree) to see some ILIAS.

The installation works and one can move inside an installation (for details what we tried, see the PR [3]). We still expect that there will be errors once one digs deeper into functionality. These will mostly be caused by us moving a lot of endpoints and resources to the new `/public` folder, which made a lot of pathes faulty. We fixed or removed many of these, but we might not have found all. Please grep for `include`, `require` and their `_once` variants in your components to anticipate according problems.

We have also moved a lot of resources like images, javascripts etc. to the public folder. This is a temporary state, we will provide according objectives that will be able to build the public folder. We will notify you once we are there. For the time being: If you modify resources that need to be delivered via HTTP, please make sure to move them into the public-folder manually. We especially moved package(-lock).json to `/public` for the moment, so make sure to call `npm clean-install` there until further notice.

We might also have missed some endpoints buried deep into the directory structure. As for the resources, we will provide according objectives to build `/public` automatically, but currently these endpoints need to be moved to the folder manually.

Other open tasks that we will tackle soon are:

* artifacts are currently not created in the desired directory.
* there is too much of node_modules in the public-folder
* for an uninstalled instance, we redirect to a file that does not exist

I would also expect you to find problems that we did not have anticipated or tasks that you consider open. Please report them via Mantis or send them to me via mail. Or, even better, try tackle them yourself and create a PR =)

Besides the public folder the most important change happened in the new components-directory. You can now start to move your components to the new directory- and namespace-structure [4] that we want to use in the future. 

When looking into components/ILIAS, you will notice that some of the components are duplicated, one with a "_"-postfix. This happened mostly when there was a folder with similar name in `/src` and `/Modules` or `/Services`. If you feel responsible for these components, please try to integrate the folder with postfix into the folder without.

As for the former point, if you have problems or questions, do not hesitate to contact me or my colleagues. If you need funding for any of the tasks you should or want to do, feel free to contact me.

Thanks y'all for moving along with us, into a future with a new and tidy directory structure and an updated component system [5] =)

Next stop will be the code for the integrations between components.

Kind regards!

[1] https://github.com/ILIAS-eLearning/ILIAS/blob/trunk/docs/development/components-and-directories.md
[2] https://github.com/ILIAS-eLearning/ILIAS/blob/trunk/docs/development/components-and-directories-process.md
[3] https://github.com/ILIAS-eLearning/ILIAS/pull/6560
[4] https://github.com/ILIAS-eLearning/ILIAS/blob/trunk/docs/development/components-and-directories.md#components-folder-and-namespacing
[5] https://docu.ilias.de/goto_docu_wiki_wpage_7295_1357.html
---
History:
* delete 'dicto' folder incl. files and adjust reference in RootFolderTest.
* move 'include' directory contents to new 'cli' directory and adjust references.
* move 'CI' directory content to new 'scripts' directory and adjust references.
* move 'sso' directory and content to 'cli' directory and adjust references.
* move 'cron.php' to 'cli' directory, delete 'cron' directory and adjust references.
* move 'server.php' and 'nusoapserver.php' from 'webservice/soap' to 'public/soap'.
* move rest of 'webservice/soap' to 'components/ILIAS/soap'.
* adjust references.
* add Update Step in 'WebService' for 'soap_wsdl_path'.
* move 'Modules' contents to 'components/ILIAS' and adjust references.
* adjust references for moved 'Modules' and include Update Steps for mail_tpl_ctx and definitions.
* move 'Services' contents to 'components/ILIAS' directory.
* adjust references for moved 'Services' and add Update Steps.
* Adjust assemblage of Exporter and ExportConfig files in  as the origin assemblage would result to wrong file names, e.g.:
nstead of it would assembly to and run into errors. Also adjust class.ilImportExportFactory => getExporterClass() => remove '_' from variable. The snake case is caused by the renaming of some Services to 'SERVICENAME_'.
* move 'xml' directory to 'components/ILIAS/Export' and adjust references.
* implement new directory structure to 'src' root folder and move files to new directories. Sort root folder 'tests' files into 'src' directories.
* adjust ILIASSuite.php for tests.
* move src directory contents to 'components/ILIAS', adjust references and move 'wopi'.
* adjust PHPUnit tests.
* move 'data' to 'cli/data' and adjust references.
* rename 'cli.php' to 'setup.php' and move 'cli' directory.
* adjust references.
* delete old 'setup.php'.
* adjust RootFolderTest and delete 'data' dir.
* adjust captainhook.json and adjust README's.
* add documentation about implementation process regarding the new components and directories structure.
* folder and file permissions and adjust ilias3.sql.
* adjust morach.json for JS Unit Tests.
* adjust data dir to public directory.
* move 'data','wopi','sso' to 'public' and adjust ilias.master.ini.php. Delete artifacts.
* move all endpoints to /public
* move inc.ilias_version.php to top-level
* adjust public directory and autoload references.
* remove include(_once)/require(_once)
* delete require_once and include_once.
* adjust required and include paths, which should not get deleted. Adjust style template paths. move delos.css, fonts, images to public/template/default.
* adjust DataFactory path.
* adjust node_modules js paths.
* adjust node_modules paths.
* adjust paths for node_modules.
* move js files from components\ILIAS directories to public dir.
* move js files from components\ILIAS directories to public dir.
* adjust js file locations.

---------

Co-authored-by: Richard Klees <richard.klees@concepts-and-training.de>
Co-authored-by: Daniel Weise <daniel.weise@concepts-and-training.de>