Exceptions: Fix `Plain Text` error handler
This commit fixes the `Plain Text` error handler. It is broken since
ILIAS overwrites the super global variables. When transforming those
back to arrays for debugging/output purposes, `NULL` bytes are introduced.
This breaks the HTTP response.
Furthermore, this commit fixes the `ilDelegationHandler` by also
returning the return value of `handle` of the current `HandlerInterface`
to the caller. The same applies for the `contentType` value provided by
the actual/wrapped handler.