UI: adjust the z-index for popover elements to avoid scrolling above … (#6940)
* UI: adjust the z-index for popover elements to avoid scrolling above the metabar. (#32769)

https://mantis.ilias.de/view.php?id=32769

This mantis ticket is solved partly. After consideration it was decided
to exchange the current webui-popover jquery library with another library in the
future, because it's not maintained anymore and hasn't been for a few
years. For this task an entry was written in the UI ROADMAP.

* UI: update the Roadmap regarding the exchange of the webui-popover library. (#32769)

(cherry picked from commit fa9ccd050a210a573d415ca7bfefe9582e8d4192)