Component: Build Public Resources
The public-folder was depopulated and the resources that should be served via the web
have been moved to the `resources/` directories of the according components. Files are
be moved to subfolders of `public/assets` according to their type, e.g. `public/assets/css`.
A common top directory for all assets will allow for various improvements regarding
delivery of the assets. Installations could, e.g., define different caching-rules for
assets than for live data delivery from other locations. Splitting up the asset folder
with regard to the type of asset we allow installations to implement effective CSP rules.

This strategy mostly panned out, besides some minor naming collisions. For two components
(COPage and the TinyMCE) the move was not possible due to the complicated structure
of how additional files are loaded from the server. These components put their assets
into specific subfolder in `public` (in components/ILIAS/COPage and node_modules/tinymce).
We expect that further moves will be possible here at some point in the future.