Use spec-aligned date format when parsing LTI score timestamp
This  changes the parsed data format for LTI score timestamps to PHP's RFC3339_EXTENDED format, which allows milliseconds and timezone offsets with colons. 

The LTI spec demands timestamps to have "ISO 8601 with a sub-second precision" format (See section 3.4.9 of the LTI Assignment & Grade Service spec (v2.0)). However, PHP's ISO8601 is not aligned with that.