Init/ErrorHandling/SOAP: Use `SOAP` error handler ALWAYS in SOAP context
This PR suggest to always use the `SOAP` error handler if
ILIAS is executed in a `SOAP` context. Currently, the `SOAP`
error handler is only used in `DEVMODE`.

After PR #7046 has been merged I thought about the error handling
again and did not find a reason why we only use this error handler
in `DEVMODE`.