CSVWriter: Remove usage of `utf8_decode`
The PHP functions `utf8_encode` and `utf8_decode` are deprecated as of PHP 8.2.
Since ILIAS 9 supports PHP 8.2 and using this functions will raise errors,
we'll have to get rid of them.

The class `ilCSVWriter` is used by several components which provide CSV exports,
but `setDoUTF8Decoding` is **not** used at all. so I removed the usage.

- BookingManager
- Poll
- Scorm2004
- ScormAicc
- Session
- StudyProgramme
- Survey
- Authentication
- Membership
- MyStaff
- Table
- Tracking
- WebServices

If approved, this has to be picked to `trunk` as well.