UI: adjust rendering of header in property forms. (#41142)
https://mantis.ilias.de/view.php?id=41142

As changes before this commit resulted to missing header content like
settings (e.g. learning progress) the rendering of the header in
property forms had to be adjusted.

If no required fields are set the header part for required text output
won't get rendered (before it rendered an empty <div>).

If at least one required field is set in a formular a required text will
get rendered in the header of the formular.