fix errors if password field is null
If users without global Admin, but full access to UserAdministration try to save datafields of accounts with admin-role, errors appear because the password field is null (instead of ""). This Commit fixes this issue. 
Since the checks in the function  "check input" all test if the password is "",  I chose to set a null password field to "" instead of adding a null-check to each if clause.