Composer: add phpunit/phpunit as dependency
This PR adds `phpunit/phpunit` as composer dependency.

Usage:
* We use it to run unit tests.
* ILIAS/Component uses its mocking framework to analyze component definitions
  during build phase.

Wrapped By:
* Not applicable, provides the binary and classes for test management.

Reasoning:
* Its the standard PHP testing framework.

Maintenance:
* Won't go away any time soon.

Links:
* Documentation: https://phpunit.de

---

This initially felt harmless, but an update to PHPUnit 10 basically destroys
our previous approach to the bundling of testcases in testsuites. Currently we
are using PHPUnits TestSuite class in various places, most notably in our global
ILIASSuite, to bundle our cases. That global suite (and some local suites) are
then added to a phpunit.xml to make them available to PHPUnit.

PHPUnit, though, seems to think of the TestSuite class as an internal class that
is used to process the organisation of tests via directories or XML. When using
PHPUnit 10 it complains that we directly add a TestSuite instance via the XML.

Hence, this uses the new order introduced by the Component Revision to allow
PHPUnit to discover tests based on an expression for directories. Individual
components could well add additional phpunit.xmls themselves to implement their
own testssuites. We could also use `--filter` to select certain tests.