MainBar/Performance: Optimize checks in `isCurrentUserAllowedToSeeCustomItem`
This PR changes the location of calling
`\ilRbacReview::assignedGlobalRoles`. Currently, it is called even
if there is no role-based visibility configuration at all.
What this PR does not address is the problem, that
`\ilRbacReview::assignedGlobalRoles`
is called for every item, which results in redundant
database requests (depending on the number of custom items).