Database/Setup: Add a hint to release "lock" state after failed database update step
If a database update step fails, the ILIAS setup
(to be concrete: the database update objectives) are in a "locked" state,
and administrators often do not know how to "release" this "lock" again
after the issue has been fixed manually or within a new Git commit.
As long as we have no option/mechanism to automatically clean up this state,
I suggest providing an SQL statement which can be manually executed.