UI: fix backwards compatibility for Data/Table backport (#7521)
The method isClientSideValueOk changing from protected to public made plugins
break that implemented their own inputs. When switching back the method to public,
unit tests fail, indicating that we cross class hierarchies when calling the
method. This was due to the fact that we indeed had two different methods
isClientSideValueOk, one from the original input hierarchy, one from a trait
for groups. These are reconciled now and thus the method can be protected again.