Fix Mantis #41498: Unable to save Anonymous Language Selection in ILIAS-Links (#7856)
* Update ilLanguage class to store language in session

In the ilLanguage class, the language setting is now stored in the session instead of in the user preferences. The condition to check the user's language preference has also been updated to exclude anonymous users. This change aims to improve language detection and application in scenarios where user preferences are not accessible or relevant.

* Fixed Copyright Issue