FileServices: correct service locator check
The check indicates that this should indeed be able to work when RBAC does not
exist in the service locator, but it actually checks an insufficient condition,
as `RBACServices` (behind `rbac()`) reaches into the service locator _again_ at
the index `rbacsystem`. The existence of `RBACServices` does not imply the
existence of `rbacsystem`, though. We could get rid of the first check by checking
and using what is actually required here (`$DIC["rbacsystem"]`).