Forum: Extract reply-prefix building code for postings
This commit extracts the code to conditionally prepend a
reply-prefix to the subject of the related parent posting
from class `ilObjForumGUI` to a separate class.
Furthermore, a unit test is added which tests different
variants of language variables defined as "reply prefix"
(e.g. "Re: "), "repeated reply prefix" and
"optimized repeated reply prefix" (e.g. "Re (4)").

The main challenge is that the prefixes are defined
in language files and can be changed in the advanced
language administration. The code has to be robust
against these changes. However, the prefix-builder
does not need to support any arbitrary translations
(shit in, shit out).

Finally: The prefix-builder should be "multibyte safe",
so it uses the `mb_*` PHP functions if available.

See: https://mantis.ilias.de/view.php?id=42322