CSS/Maps: 42840, restore missing styling for OpenLayersMap and adjust references to the pointer.
Include a the new icon namend "icon_mapm.svg" (to be found in "components/ILIAS/UI/resources/images/standard") for maps because the original icon still was a PNG file. This was decided by the Iconset Task Fource on the 2th of december 2024.
The icons "mm_20_blue.png" and "mm_20_shadow.png" were deleted as they are not needed anymore.

https://mantis.ilias.de/view.php?id=42840