External link behavior for same-system URLs in ILIAS
External links in ILIAS always open in a new tab, even if they point back to the same system. Ideally, this could be avoided by properly setting internal links, but this is not consistently the case throughout the platform. In particular, we cannot change the link to the login page.

This commit adds a JavaScript snippet that rewrites "external" links pointing to the same system, ensuring they behave like internal links.