Portfolio: Remove `back_url` usages
This commit suggests removing all `back_url` related code
in the `Portfolio` component.
First of all: Building links with a named `back_url` HTTP query
parameter in `\ilObjPortfolio::getAvailablePortfolioLinksForUserIds(...)`
does not really work. The "URL Builder" does not support
any kind of "named" query parameter in the generated URLs. A value
in the `$a_params` array is just added as a path segment,
e.g.: "go/.../[BACK_URL]".

Furthermore, the permanent link handling in `\ilObjPortfolioGUI::_goto(...)`
expects `$id[1]` to be of type `int`. I came to this conclusion
after checking the usages of `user_page`.
The current code results in the "Back URL" being passed to query
parameter `user_page`, which then results in a `Refinery`
error in the subsequent HTTP redirect:

`"The value "'[URL]'" can not be transformed into an integer"`

I then checked all usages of `back_url` in the `Portfolio` and
identified, that there actually no real usage. All PHP variables
built/assigned with values from
`ILIAS\Portfolio\StandardGUIRequest::getBackUrl(...)` are never used:
- The code in `ilObjPortfolioBaseGUI::__construct(...)` only appends
the value to URLs built with `ilCtrlInterface` in the current request.
- The code in `ilObjPortfolioBaseGUI::preview(...)` does not use
  the build values `$back` and `$back_caption` at all.

Because of this, I also removed the `$a_back_url` from the callers
of `\ilObjPortfolio::getAvailablePortfolioLinksForUserIds(...)`.