fix: error with phpunit import and some fixes in ilInitialisation
When administrators execute `composer install --no-dev` they get an error related with Reader.php. This error was reported in [Mantis](https://mantis.ilias.de/view.php?id=41248) and it is caused for an in import. Composer cannot find MockBuilder because the dependency phpunit is `require-dev` instead of in `require`.
In the other hand, I found some errors that do not allow ILIAS to initialise using `ilInitialisation::initILIAS();`. With theses changes we can fix those errors.