fix: LTI avoid duplicates in User Management
With this change, we fix the ticket [0041165](https://mantis.ilias.de/view.php?id=41165)

The suggested solution could not be fully implemented because the duplicate names shown in the image are the names of LTI providers (those configured at the platform level), which were displayed once for each resource that used them to provide content. Therefore, instead of the suggested change, we will now display the name of the provider together with the name of the resource it provides in order to avoid confusion.