fix: save phpSessionId after authentication
In ILIAS 10, the community found an error related with LTI connections between platforms allocated in different domains. In these cases ILIAS wasn't able to connect to another ILIAS platform if they were hosted in different domains. The reason was the privacy policy of the browsers. For that reason we set the session's cookie during the authentication in ilStartUpGUI. The error was found in ILIAS 9, for that reason we propose the same change for ILIAS 9.

In ILIAS 10, the PR was https://github.com/ILIAS-eLearning/ILIAS/pull/9179