fix: redirection after terms of service
When the provider platform has terms of service the consumer is redirected to the page to accept or not this terms, but when the consumer accept them ILIAS is not able to arrive to the correct shared object. This is because LTI doesn't save the original target. For that reason we propose save that original direction when the user logs in. The same change was introduced in ILIAS 10 in the PR: https://github.com/ILIAS-eLearning/ILIAS/pull/9250

Related with the report [0039629](https://mantis.ilias.de/view.php?id=39629)