ApacheAuth: Ignore `force_login` command
This commit suggests ignoring the `force_login`
command in the `tryAuthenticationOnLoginPage` function.

If the public area is **enabled**, and ILIAS detects
that the user has **no valid ILIAS session and access**
to the requested resource, an HTTP redirect to the
login view with a `cmd=force_login` query parameter
will be initiated. This currently leads to a problem
where the automatically initiated "Apache Authentication"
Single Sign-On will **not** be triggered, even if configured
this way.

Effect: With this change, requesting the login page with
an **enabled** "Apache Authentication" will only be possible
**without** triggering the automatically initiated Single Sign-On,
if `passed_sso=1` is given in the query parameters (no change was
required to achieve this behaviour).